requirements.txt
django
django-extensions
django-bootstrap3
psycopg2
workalendar
jquery
djangorestframework

